prizes simplified code, spacing, styling which is overlap with jumpstart)
