--- src/components/GirlsHooHack/Prizes/PrizeItemDesktop.tsx
@@ -1,18 +1,0 @@
-import { Col } from 'react-bootstrap';
-
-interface Props {
-  key: string;
-  name: string;
-  detail: string;
-}
-
-function PrizeItemDesktop(props: Props) {
-  return (
-    <Col xs={4} className="hack">
-      <h4>{props.name}</h4>
-      <p>{props.detail}</p>
-    </Col>
-  );
-}
-
-export default PrizeItemDesktop;
