+++ src/components/GirlsHooHack/Prizes/Prizes.tsx
@@ -2,8 +2,7 @@ import { useEffect, useState } from 'react';
 import { request } from 'graphql-request';
 import { Col, Row } from 'react-bootstrap';
 import { isMobile } from 'react-device-detect';
-import PrizeItemMobile from './PrizeItemMobile';
-import PrizeItemDesktop from './PrizeItemDesktop';
+import PrizeItem from './PrizeItem';
 
 function Prizes() {
   const [prizes, setPrizes] = useState([{'id': '', 'prizeName': '', 'description': ''}]);
@@ -22,10 +21,8 @@ function Prizes() {
           }
         `
       );
-    
       setPrizes(hackathonPrizes);
     };
-
     fetchPrizes();
   }, []);
 
@@ -33,21 +30,22 @@ function Prizes() {
   <Row className={`bg-white justify-content-center py-5`}>
     <Col sm={10} xs={11} className={isMobile ? "my-3" : "py-5"}>
       <Row className={isMobile ? "" : "my-5"}>
-        {isMobile ? <Col>
-          <h2 className="text-orange hack">PRIZES</h2>
-          
-          {prizes.map((prize) => 
-            <PrizeItemMobile key={prize.id} name={prize.prizeName} detail={prize.description} ></PrizeItemMobile>)}
-          
-        </Col> 
-        : 
         <Col>
-          <h2 className="text-orange hack">Prizes</h2>
-          <Row>
-            {prizes.map((prize) => 
-              <PrizeItemDesktop key={prize.id} name={prize.prizeName} detail={prize.description} ></PrizeItemDesktop>)}
-          </Row>
-        </Col>}
+          {isMobile ? 
+              <div>
+                <h2 className="text-orange hack">PRIZES</h2>
+                {prizes.map((prize) => 
+                <PrizeItem key={prize.id} name={prize.prizeName} detail={prize.description} ></PrizeItem>)}
+              </div>
+              :
+              <div>
+                <h2 className="text-orange hack pb-3">Prizes</h2>
+                <Row>
+                  {prizes.map((prize) => 
+                    <PrizeItem key={prize.id} name={prize.prizeName} detail={prize.description} ></PrizeItem>)}
+                </Row>
+              </div>}
+        </Col>
       </Row>
     </Col>
   </Row>
