+++ src/index.scss
@@ -116,6 +116,11 @@ h3.hack {
   font-weight: 700;
   line-height: 42px;
 }
+h4.hack {
+  font-size: 22px;
+  font-weight: 700;
+  line-height: 29px;
+}
 p.hack {
   font-size: 18px;
   font-weight: 300;
@@ -188,6 +193,10 @@ p.hack {
   h3.hack {
     font-size: 20px;
     line-height: 26px;
+  }
+  h4.hack {
+    font-size: 16px;
+    line-height: 21px;
   }
   p.hack {
     font-size: 12px;
